Refactor sds to allow sds-ee and sds-ee-fips to be created (#9065)
* Move sds "main" code out of "main"

Update _helpers.tpl

Update _helpers.tpl

* Force a pull

* rename file with moved "main"

* sds-fips helm test

* Update refactor-sds-project.yaml

* Update helm_test.go

---------
